# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MaintSys API",
        "description": "API de gerenciamento de manutenção industrial - MaintSys",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/maintenance-logs": {
            "get": {
                "tags": [
                    "Logs de Manutenção"
                ],
                "summary": "Lista todos os logs de manutenção",
                "operationId": "3f1612f2c1dafe6278765537b0f75529",
                "responses": {
                    "200": {
                        "description": "Lista de logs de manutenção",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "machine_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "service_order_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "action": {
                                                "type": "string",
                                                "example": "Troca de óleo"
                                            },
                                            "defect_type": {
                                                "type": "string",
                                                "example": "Desgaste"
                                            },
                                            "machine": {
                                                "type": "object"
                                            },
                                            "serviceOrder": {
                                                "type": "object"
                                            },
                                            "user": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Logs de Manutenção"
                ],
                "summary": "Registra um novo log de manutenção",
                "operationId": "c16988cb3995dc8454458a8025ab74c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "machine_id",
                                    "action"
                                ],
                                "properties": {
                                    "machine_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "action": {
                                        "type": "string",
                                        "example": "Substituição de correia"
                                    },
                                    "defect_type": {
                                        "type": "string",
                                        "example": "Quebra mecânica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Log de manutenção criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "action": {
                                            "type": "string",
                                            "example": "Substituição de correia"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "machine": {
                                            "type": "object"
                                        },
                                        "serviceOrder": {
                                            "type": "object"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/maintenance-logs/{id}": {
            "get": {
                "tags": [
                    "Logs de Manutenção"
                ],
                "summary": "Retorna um log de manutenção específico",
                "operationId": "cfb11ef88102c1c0f804124665a63251",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do log de manutenção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do log de manutenção",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "action": {
                                            "type": "string",
                                            "example": "Troca de óleo"
                                        },
                                        "defect_type": {
                                            "type": "string",
                                            "example": "Desgaste"
                                        },
                                        "machine": {
                                            "type": "object"
                                        },
                                        "serviceOrder": {
                                            "type": "object"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Log não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Logs de Manutenção"
                ],
                "summary": "Atualiza um log de manutenção",
                "operationId": "04fa9aade5e79a33bf197d64d5dc9e97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do log de manutenção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "example": "Lubrificação revisada"
                                    },
                                    "defect_type": {
                                        "type": "string",
                                        "example": "Folga"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Log de manutenção atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "action": {
                                            "type": "string",
                                            "example": "Lubrificação revisada"
                                        },
                                        "defect_type": {
                                            "type": "string",
                                            "example": "Folga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Log não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Logs de Manutenção"
                ],
                "summary": "Remove um log de manutenção",
                "operationId": "e5256e91503c091ad1b1a38e8a833af6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do log de manutenção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Log removido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Log não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}